hostapd_cli: support disable and enable commands on mld iface
Currently, hostapd_cli's enable/disable commands enables or
disables all the bss on a specific interface on which the
command is given. Extend this support for ifaces which has
MLDs as well. Since dynamic deletion or addition of a MLD
link is not supported, delete the whole MLD when the bss of
a link is getting deleted. This recursively deletes all the
bss of the partner link as well. hapd_iface which are not
partner to any of the MLD of the disabled link has no
impact.

If all the remaining cohosted bss (which are not partner
to the MLD in the deleted link) should not be impacted, the
suggestion is to run these bss as a separate hostapd instance.

With MBSSID case, even the current design is removal of one of
the Tx/Non-Tx vap will remove all the vaps in the group.

There is no change in command/args. Disable or enable
of any link bss will disable/enable all the bsses in that
hapd_iface as well as other hapd_ifaces which are partners
of the MLDs in the disabled/enabled iface.

Change-Id: I17a27e9174bd9e0a8f4ba0dd3613de3cbecaf93f
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>