hostapd: fix bandwidth reduction in 320 MHz DFS case
Currently, once DFS interference is detected and bandiwdth is
reduced, the bandwidth change is already altered in the iface
config. The function hostapd_dfs_request_channel_switch() expects
the original bandwidth in its argument since the new reduced
bandwidth it can take from the the config. However, at present,
the new bandwidth is passed which is wrong.

Fix this issue by passing the original bandwidth.

Fixes: 0caacea566d5 ("hostapd: handle DFS detect event during MLO operation")
Change-Id: Iecb3fda4ff69434fb28cd6646051f63344c8c855
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>