hostpad: fix advertising DS parameter for legacy (11b) mode
Currently, DS parameter is only advertised for 11g mode. However, 11b
is also in 2.4 GHz and hence DS parameter should be advertised for it
as well. Due to this, few clients are not able to scan the beacon itself
when AP is in legacy 11b mode and hence can't connect to it.

Fix this issue by advertising DS parameter for 11b mode as well.

Fixes: 6be45acafd88 ("hostapd : fix DS parameter in beacon and probe response")
Change-Id: I51bac27103e740bd8b808ed2eecc2a1a3a717cab
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>