hostapd.sh: Add frequency list for sta mode
Currently, in single wiphy architecture, when sta interface is brought up
in one particular band from OpenWrt config file, then it scans for
all the bands instead of one.

Because of this, when same ssid is broadcasted on 2G Band and 5G band,
then sta selects the 2G band.

In order to overcome the above issue, add the frequency list of the band to
the wpa_supplicant config file. Hence, it would scan only the configured band.

Change-Id: I51eb1f0cee159599af1d7421d237edfc7e7b0aa3
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>