hostapd: add Max Channel Switch Time IE support
The Max Channel Switch Time (MCST) element indicates the time delta
between the time the last beacon is transmitted by the AP in the
current channel and the expected time of the first beacon transmitted
by the AP in the new channel.

IEEE Std 802.11be-Draft2.2-2022, Section 35.3.11 (Multi-link procedures
for channel switching, extended channel switching, and channel quieting)
indicates that if an AP affiliated with an AP MLD is switching channel,
MCST IE shall be included in every beacon and probe response frames it
transmits.

Add support to include MCST IE in beacon and probe response during channel
switch announcement for an AP affiliated to an AP MLD.

Change-Id: I70ce171ea2aa998ceeaaeaa4539bb0191ed55282
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>