hostapd: Add bcast deauth before deinit
The AP has to ideally send deauth frames to notify
the associated STA about disconnection. If the AP
fails to send deauth and goes down, STA still thinks
it is associated with AP, until it sees a beacon miss.

Adding this change makes the process little more efficient
as we cannot rely only on STA to notice the missed beacon

Change-Id: Ic60e82918240dab6f971eba029f6bca0bdd954ae
Signed-off-by: Ramanathan Choodamani <quic_rchoodam@quicinc.com>