hostapd: fix 5 GHz/320 MHz channel switch
During interface bring up, in case of 320 MHz, only 12 sub-channels
are considered to check whether CAC is required or not. However, during
channel switch, all 16 channels are checked which is wrong. This also leads
to CSA failing and if incorrectly CAC required is detected, interfaces will
be disabled and enabled again which is also wrong since CSA should have
happened seamlessly in that case.

Fix the above issue by considering 12 sub-channels only in case of channel
switch to 320 MHz as well like we do for bring up time.

Fixes: 7c8c82d240b1 ("hostapd: Add 5GHz 240MHz support")

Change-Id: I6734467a133070c2e65d524665bb31b076b1d021
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>