hostapd: fix null pointer access in acs_retry
In repeater scenario, when AP interface is configured
for auto channel and STA interface is doing scan. AP interface
schedules acs scan retry as STA is currently doing scan.

In this meantime, AP interface is disabled using hostapd_cli and
later, when the scheduled acs scan is triggered it is crashed
with null pointer access while using iface details because
it was cleared during interface disable.

To handle this, cancel the scheduled acs scan retry while disabling
the ap interface in acs_cleanup.

Fixes: I5a83170a06d (hostapd: WAR for ACS scan)

Change-Id: I8f7d8d9ddadcf6e7995a27b78e6d642161451e71
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>