hostapd: Fix channel switch when STA is not associated
Currently, when STA is not associated and channel switch command is given second
time, output is FAIL. This is because in function mlme_event_ch_switch we are
not able to send channel switch start and completion event. This issue is seen
in MLO.

Fix this issue by updating condition to check if interface is connected or not.

Fixes: Ib54f0c13c7f3 ("Upgrade hostapd package until commit a9012070a (Feb 21, 2023)")

Change-Id: I22e9996cbcb4f753b57fe996654d38737b3591e3
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>