hostapd: fix NULL pointer access in i802_init()
Currently, i802_init() function assumes that always bssid
will be provided by the user via the config. And directly
uses it in memcpy() function. If user has not provided it,
then this leads to NULL pointer access and hostapd crashes
with "Segmentation Fault" error.

It is worth mentioning that, passing bssid in config is not
mandatory as per the documentation hence there could be such
cases. For example - Quick Track Certification Test case.

Fix this issue by removing the part where it is blindly assumed
that bssid will be present.

Fixes: 0c04ec4caf43 ("hostapd: ml interface bring up changes")

Change-Id: Ie06ee6259c3613bb0c5f6a56794fb2cb7aac3ecf
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>