hostapd: dfs puncturing not applied when radar detected during cac
When radar is detected during a CAC, it discards the CAC and
determines another available channel to operate instead of operating
configured channel with puncture affected sub channels.

Fix the issue, when radar detected during CAC, the user space application
validate the radar bitmap, if it is matches valid puncturing pattern,
it will restart the CAC with valid pattern to puncture radar affected
sub channels.

Change-Id: Ib0eb9e293941f37fd47173479f9efd1181a3c7ec
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>