hostapd: fix CSA with puncturing pattern
Currently, if due to puncturing pattern, final width comes out to be 20 MHz,
secondary channel offset is not updated in the settings structure. Due to
this, width is taken as 40 MHz since enum for width is 0 for both 20 and
40 MHz. Using secondary channel offset, it is differentiated and in issue
case, offset was not reset.

Because of width taken as 40 MHz, Wide Band Channel Switch (WBCS) Wrapper IE
is also added in the beacon. However, when there is movement to 20 MHz, WBCS
wrapper IE should not be included or else this will lead to clients
connectivity issue.

Fix this issue by updating the secondary channel offset value properly in
the structure.

Fixes: a2318a8dc4ac ("hostapd: Puncturing support for interference affected dfs sub channels")
Change-Id: I0462567578a1e23c54b9e46949e5dcd51ae6fcfc
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>