hostapd: fix NULL pointer access in update_pmksa() function
update_pmksa() function initialises the ML wpa but before that it ensures
that all partners for the give BSS has the same pmksa reference. There is
a possibility that the partner BSS has not been set yet (due to CAC) and
hence the reference will not be set. In this case, directly accessing the
reference leads to NULL pointer access and hostapd crashes.

Fix the above issue by checking whether first the reference exists or not
for the iterating link BSS.

Fixes: 4ef2714337af ("hostapd: WPA2 and WPA3 Secure mode changes for Multi link AP")
Change-Id: Id3bdcd19b95bd3eaa65057648564092f0c684fe8
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>