hostapd: fix premature beacon set during association handling
Currently, during association handling, if any of the iface parameters
changes due to some reason, it calls ieee802_11_set_beacons() function.
This function sets beacon for the all the BSSes in that iface even if
the beacon was not set already. This leads to setting the beacon
prematurely for some BSSes which was intentionally not started.

Fix the above issue by calling ieee802_11_update_beacons() function
instead. This function only updates the beacon if it was already set.

Change-Id: I65fe4680b2420d486f9d06351411c214f12cc3f0
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>