hostapd: fix link BSS de-initialize
Currently if first link in a MLD is disabled, it makes the next link
as the first link. However, other link's first BSS pointer is not
updated which is wrong. This could lead to issues when other link
tries to access the first link BSS since it will still point to older
first link which is disabled. It should point to the new first BSS.

Fix the above issue by updating the first BSS pointer of all the links
in the MLD.

While at it, fix caching the scan link info as well [minor]

Fixes: df4995a8b6d3 ("hostpad: Rework MLO support based on upstream design")
Change-Id: I2eaffc02782080692e01e225f4dbfc4e2ec6cf44
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>