hostapd: Fetch correct channel width for HE operation IE
When AP with 6GHZ EHT320 configuration comes up, 6GHz operation Information
field in HE operation IE is updated with wrong ccfs0 and ccfs1 value.
Third party clients is not able to scan the beacon and update the scan
list due to it.

Fetch the data from he_oper_chwidth for HE operation IE.

Change-Id: Ic2ea87273e4272133c70a86d0a7b965ea89cc726
Signed-off-by: Karthik M <quic_karm@quicinc.com>