wpa_supplicant: Add channel 140 to htplus allowed list
When channel 140 is configured in mesh, interface fails to come up
due to channel bond (136,140). Channel 136 is not HT40+ capable and hence
validation for HT channel bonding fails.

In mesh, during channel setup, secondary channel offset for the
configured channel will be selected as +1 if primary channel is capable of HT40+.
In current code, channel 140 is not allowed as HT40+ and hence secondary
channel offset is selected as -1, 136 is selected as secondary channel.
But channel 136 is not HT40+ supported and fails in channel bonding
validation.

Fix this by adding 140 to HT40+ allowed list because it is capable of
HT40+

Change-Id: Id3799ebbe4c59bfb16b607614f0b7583c213b7d8
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>