hostapd: Add support for SW Progressive DFS
When RADAR is detected in the opearing bandwidth (OBW), device is
required to vacate the affected channels either by reducing the
bandwidth or by completely vacating the OBW. SW based progressive DFS
recipe allows the device to continue radar detection on the affected
channels by setting the device bandwidth (DBW) same as before while
reducing the OBW to avoid the affected channels. Tx will be restricted
to OBW while DBW will be used for Rx.

Change-Id: I57c087fe731c9339f978079d240d305bdda8b898
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>