hostapd: fix advertisement of MLD capabilities
Currently, hostapd directly advertises the MLD capabilites received from driver.
Since this info is exchanged during init time only, driver will advertise the max
supported possible values. Hostapd should parse it and then based on current
situation should fill the values accordingly.

For example, Max number of simultaneous links is supposed to be value between 0
and 14, which is the number of affiliated APs minus 1. Driver advertises this
value as 5 and hostapd, irrespective of current active links, it puts 5 in the
frames.

Fix this issue by parsing the value and at the same time using the values as
per the current situation of the links. The advertised values will be used
as the upper limit.

While at it, also print the MLD capabilities in debug print.

Change-Id: Iafeacb202b957707b79184d073addc4c91b5013b
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>