hostapd: fix BSS deinitialization
Cuurently, if in an iface, first interface is in single link and
next BSS is in multi-link where it is not the first BSS, then
during deinitialization, since for the MLD interface, interface
created flag would not be set, it wont even remove the link from
the interface which is wrong.

Also, during disable, if the iface was not the first one
during the hostapd invoking command then the first interface in
that iface would be deleted. However, the interface needs to be
retained since if enable is given back, hostapd needs the first
interface of an iface to be already present in order to start
its init process.

This fixes the above two issues.

Fixes: ee5f02113cf7 ("hostapd: MLO: optimize MLO flows")
Change-Id: I41974d857303fcbcff05e7c7e9a39c8b54b16a27
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>