hostapd: fix ML Link ID in ML RNR formation
Change for fixing BPCC counter issue assumed match_idx would be non-zero
for self-partners which is wrong. For self partners it would be 0, hence
code was going to false condition and incorrectly filling Link ID resulting
in ML association issue.

Also, by mistake it updated mld_update param to true which results in co-located
RNR issue. Remove that as well.

Fixes: 975533432bea ("hostapd: fix BPCC count value in ML RNR")
Change-Id: I28fb8d0eccd260f31b28163c251b246518df978e
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>