hostapd: fix wpa_auth conf during reconfig
Currently when wpa_auth is reconfigured, wpa_auth conf is generated
and directly used. This leads to wpa_auth trying to set up its state
machine in such a way that it is beyond the driver supported
capabilities hence few requests to kernel are rejected leading to its
state machine entering failure state. Hence no client can connect to
it once it is reconfigured.

However, in normal bring up sequence, after the conf is generated,
it is sanitized with the iface capabilities. Hence, initial wpa_auth
state machine setup is successful since it never tries to go beyond
the supported capabilities.

Fix the above issue by moving the sanitizing the conf part to a
helper function and calling it during initial as well as during
reconfig time.

Change-Id: If6b8f63b2f09c83af1eb0026de5117e3804615f5
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>