[CAF Only] hostapd: Fix VHT center frequency for EHT 240 MHz
Currently, when 240 MHz is configured in 5 GHz band, the VHT center
frequency is not set correctly.

This is because in hostapd_set_oper_centr_freq_seg0_idx, the
center frequency is not calculated for 5 GHz 240 MHz case. So, it's
frequency is set the same as 320MHz.

When radar is detected, the punct_update_legacy_bw receives bandwidth
of 160 MHz but center frequency for 320 MHz. This causes wrong
calculation of new channel bandwidth during puncturing.

Fix this issue by setting the correct center freqency and getting
it from hostapd_get_oper_centr_freq_seg0_idx instead of directly
giving VHT frequency.

Change-Id: I564e74b915f26e059e198801787bbe0a6a518084
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>