hostapd: fix RNR building for MLO
When forming RNR element for MLO case, conditional check is there to
avoid inclusion of own interface report by validating the channel
number. Due to this check, if same channel number is set on both
5 GHz and 6 GHz interfaces, then partner 5GHz/6GHz link neighbour
report is not getting added on RNR element and as a result association
is not happening on all the links.

Fix this by checking frequency instead of channel number.

Change-Id: I4cadd6e3a51248429ae5e3600ed2ce02598d0458
Signed-off-by: Rathees Kumar R Chinannan <quic_rrchinan@quicinc.com>