hostapd: clean up MLD when not required any further
Currently, whenever a new BSS is created, if it is an EHT then it is tied
to a corresponding MLD structure. If structure does not exist already, a
new one is created and tied to it. Accordingly link ID is assigned as well.
However, when BSS is deleted, the MLD structure is not freed and when the
next time it is again created, it increments the link ID further and gets
a wrong link ID.

For example - 2 GHz SLO AP case.
First Wifi, link ID 0 would be assigned and mld wlan0 would be created.
When wifi down is issued, BSS would be deleted but mld wlan0 will not. When
again wifi is issued, BSS will tie back to mld wlan0 but this time link ID
will be incremented again and 1 would be assigned. Hence at subsequent wifi
down/up, link ID keeps on incrementing.

Since link ID remains till lifetime of the BSS and MLD, the next link ID
counter can not be just reset back to 0 when a BSS is deleted. Otherwise,
in interleaved link enable/disable case, link ID would be changed.

Hence to overcome this situation, add changes so that whenever a BSS is
deleted, if the MLD is not referenced by any other existing BSS, delete
the MLD structure itself.

In order to know how many BSSes are referring a given MLD, introduce a new
member refcount in MLD. If he value is 0 then it is safe to delete the MLD.

Change-Id: I27380c619db9e950dba2d17ed551f58e14ab3cc6
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
(cherry picked from commit ec0beb32b73e39541fbbde406609d568fa59c3cd)