hostapd: fix the check for device BW parameters
Device BW parameters will be set to default 0 if not sent by kernel.
This is resulting in a regression where the AP does not change the
channel upon radar detection because hostapd_dfs_radar_detected()
compares only the operating and device bandwidth values to decide
the present of device bandwidth and modifies the operating bitmap.
Call hostapd_is_device_params_present() instead.
This function is already called by hostapd_dfs_radar_detected() once
so cache the result locally instead of calling it twice.
Replace incorrect bitwise OR operations with AND operations.

Fixes: I57c087fe731 ("hostapd: Add support for SW Progressive DFS")
Change-Id: Ib35343cf2e0dc42f5a29d40f78d54671ce343b44
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>