hostapd: fix 5 GHz frequency boundary check
Condition to check if frequency lies outside the 5 GHz range is wrong in
handling the radar events. Due to this, for non 5 GHz link also, the DFS
events are routed and later channel comes out to be NULL and if accessed,
it leads to seg fault.

Frequency can not be lower and greater than two different boundaries at
the same time. Hence using logical AND in the condition is wrong.

Fix this by using logical OR instead.

Fixes: dea791c6b476 ("hostapd: Add support for SW Progressive DFS")
Change-Id: I818f19037284f625cb83e5a52fb6d63ef17bb99d
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
(cherry picked from commit 8f0277ff17988bab5a238cefb2b0b35bff50ea5a)