hostapd: fix overwriting MLD info during auth handling
Currently whenever we handle auth frame and if the transaction is one, we
memset the mld info and read it again with the incoming frame. However,
this should be done only when kernel is also aware of the change in the
sta info. Or else, source address will be STA's MLD MAC address and it
will be copied in link address which will be wrong.

This leads to issue where in STA entry is created with its link address as
MLD MAC address and hence even after address translation, frames are sent
out on MLD MAC address which is incorrect.

Fix this by memseting the MLD info and processing it only when the STA is
going to get re-added in the kernel so that the info is same across hostapd
and the kernel.

Change-Id: If74364b82c98ae6a821564e803db5d9dc5f54263
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
(cherry picked from commit 3b2c7f4e5ea358277266248ac0c9d1b62324c84d)