hostapd: Disable AP if invalid parameters are provided in CSA
Existing code does not disable AP when incorrect
bandwidth and frequency parameters are used in
channel switch command. At the start of the function
hostapd_ctrl_iface_chan_switch(), if disable_csa_dfs is set
hostapd cancels the timeout even before checking the
validity of provided channel switch parameters. So if incorrect
parameters are used in the existing channel switch
command the AP will continue to function with the
current parameters. This is against the DFS rules as
AP keeps operating in the radar affected channels.
To fix this, the existing timeout cancellation is
removed from the function start. The
hostapd_dfs_radar_handling_timeout will be cancelled
once the frequency parameters are validated and before
channel switch parameters are sent to kernel.

Change-Id: Id7f12e7d607210fe7cc524c706a9978d16e08ea0
Signed-off-by: Shivani Tambatkar <quic_stambatk@quicinc.com>