hostapd: Update the hostapd module to handle the NULL value of wds_bridge.
By default, the wds_bridge parameter in the hostapd config file
is NULL character,So it updates the NULL character as bridge name.
Because of this, it fails to add the wds interface into the
bridge while running the  hostapd manually.

This change checks the wds_bridge parameter and assign
the wds bridge name as per in the hostapd config, when the
character is not  NULL. Otherwise, the wds_bridge remains
in the default bridge interface name.

Fixes: If5d08c2d23 ("hostpad: Rework MLO support based on upstream design")

Change-Id: I08c3dbfa5b0992385a3cc7f43c4805a3804f1c79
Signed-off-by: Arulanbu Balusamy <quic_abalusam@quicinc.com>