hostapd: fix 6 GHz centre frequency in AWGN interference
Currently, when AWGN interference is detected in the primary channel
of 6 GHz band operating in 320 MHz and disabled channels are present
in the first 160 MHz of the 6 GHz band, channel switch fails.

This is because the bandwidth for which centre frequency is
calculated and the bandwidth for which disabled channels check is
done are not the same.

Hence, centre frequency is selected for channel switch even when
disabled channels are present in that bandwidth range.

Fix this issue by updating the correct centre frequency for the
bandwidth.

Fixes: I6a9547417b6 (hostapd : add awgn support for 320 MHz bw in 6 GHz
band)

Change-Id: I3cad5eba6fbee5186595e1805a54d99e2725f0f7
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>