hostapd: WAR for rtnl socket allocation failure
Currently, FT roaming on secondary links in MLD is
not happening when station tries roaming.

When a MLD is created, the hostapd will add the link
addresses to the bridge through rtnl socket. After package
upgrade, rtnl socket allocation is not happening as the bridge
is busy.

Adding WAR by checking the bridge index for rtnl socket creation as the netifd
will add all the interfaces created with "lan" configuration to br-lan

Change-Id: I2cef3d1afaa34e2b21492c61b9e470df50190434
Signed-off-by: Karthik M <quic_karm@quicinc.com>