hostapd: Set eht operation tag's basic mcs nss value.
Currently the ext tag: eht oper's basic mcs nss value is set to the
max tx and rx nss values. This is an issue because if the station's
rx and tx nss are below the basic mcs nss of the eht oper in AP, the
association will not happen in EHT mode(as per draft doc at EOM).

Hence the basic mcs nss is now set to be 0x00000011 instead of the max
nss values, so that the association between station and AP can now
be done with the EHT mode.

w.r.t. the draft doc.: IEEE 802.11-23/0638r4 "The basic EHT-MCS and
NSS set is the set of <EHT-MCS, NSS> tuples that are supported by all
EHT STAs that are members of an EHT BSS"

Change-Id: Iff7d89eb807da3e738b4799b25f9d315283c6f1e
Signed-off-by: Naveen S <quic_naves@quicinc.com>