[WAR] hostapd: fix drv corruption due to choosing invalid wiphy_data
After insmod, the pointers of older wiphy_data pointing to drv are
expected to be removed. But here the pointers of the older wiphy_data
are still pointing to drv which is resulting in drv corruption.

This fix will remove the links pointing to drv from the older wiphy_data.

Change-Id: Idd40eaa4db9941d89e3dd82e85e9357bf1917fa7
Signed-off-by: Monika Korada <quic_koramoni@quicinc.com>