hostapd: Fix CU beacon delay in channel switch
Currently, when channel switch is triggered in AP, critical update
flag in the new channel beacon is set after a delay.

This is because, the CUF for elements modified after channel switch
is not set immediately after the channel switch completes. Instead,
this flag is set after beaconing starts in new channel.
- First, CUF for elements added is set during the Channel Switch
- Channel Switch completes and beaconing starts immediately.
- Next, CUF for elements modified is received when set beacon is
called.
So after CS is finished, beaconing starts immediately and the
CU modified flag is received later.

Fix this issue by adding a CUF after csa completes and unsetting
critical update during set beacon.

Change-Id: I1c10e3164cc53bbed04672420c71656c4cef0918
Signed-off-by: Mohan Kumar G <quic_mkumarg@quicinc.com>