hostapd: fix channel switch on AWGN interference
Currently when AWGN interference is detected in the primary
channel of 6 GHz band operating in 320 MHz and configured for
AU region, channel switch fails

This is because intf_awgn_find_channel_list() function will
try to find a new channel with same or nearest lower possible
bandwidth which has no interference. The logic used to check
whether the selected channel has interference in it has a
problem. It assumes entire frequencies in the current channel
has interference in it and if the newly selected channel has
a frequency common with the current channel (overlapped), then
the current logic will not use the new channel. So the frequencies
from current channel which has no interference can't be reused
by this logic.

This causes problem when bring up 6G AP for AU region with 320MHz
bandwidth. There is only one channel available for this configuration
and it will use all the available frequencies. So if any
interference happens in the current channel there is no other
frequencies available for a new channel.

Fix this issue by reusing frequencies from current channel with
no interference by allowing new channel with overlapping frequencies
with current channel but discard them only if it includes
frequencies with interference

Change-Id: I38b400de878b7606faa603000c13c286909967dc
Signed-off-by: Ajith C <quic_ajithc@quicinc.com>