hostapd: fix NULL drv pointer access
Currently drv will point to wiphy_data when mld is created.
When channel switch is done, the older wiphy_data will be
removed and added again. When the older wiphy_data is removed,
the pointer pointing to drv will be removed and it will not be
pointing to anywhere.

When mlds are added again, drv needs to point to new wiphy_data.
But currently drv is pointing to NULL and when we try to access
the NULL pointer it results in hostapd crash and segmentation fault.

This fix will make drv to point to itself after the pointers of old
wiphy_data pointing to drv has been removed.

Change-Id: Ic1339a926f5c0a5b3828eb988ef73ed4b598dd98
Signed-off-by: Dhanavandhana Kannan <quic_dhanavan@quicinc.com>