hostapd: fix HE Operation IE seg0 and width during puncturing
Currently during puncturing case, the width and seg0 displayed in HE
Operation IE is incorrect. This is because the helper function which
calculates the new width and seg0 expects the original values with
respect to EHT mode but however HE mode values are passed which results
in wrong calculation.

Fix this by setting the initial value before calling the helper function
with the values with respect to EHT mode.

Fixes: ade3d3aa15ef ("hostapd: update proper 6 GHz channel information in HE Operation IE")

Change-Id: Ie5bc7b9047e1726fbc01f9b6641a8b9ae7786f4a
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
(cherry picked from commit 8e59d21cbecf08795f54f334ff8ac23cc950bf05)