hostapd: Add FT over DS support as per D4.0 spec
In case of FT over DS in MLO, the FT req from a STA can come from one
link and the assoc/reassoc req can come from another. In order to handle
this situation, D4.0 spec mandates STA's ML address instead of link
address in the FT req.

On receiving the FT req from the STA by the source AP, add STA's ML address
in  FT-RRB request frame to the target AP. Since the target AP is not aware
about which link hapd the STA should be added, use a look-up table to store
the sta info till the reassoc request is received. Store the R1KH in
wpa-state-machine so that it can be used later instead of fetching it from
config. This is needed in the FT-over-DS case where the STA can use
different links for FT-request and reassoc-request. After adding the STA
entry to the look-up table, send FT-RRB response back to the source AP
which forwards it to the STA.

STA then sends a reassoc request to the target AP. On receiving the
reassoc req from the STA, fetch the STA entry from the look-up table
and add it to the hapd of the link in which the reassoc req was received.
Add a flag in sta_info to notify driver about the auth_alg of the STA.
This is used to prevent the FW from waiting for the 4-way handshake
before opening the port for transmission.

Change-Id: I2b6637cdd0339935c64e80c71803c5d1bbffcf27
Signed-off-by: Amith A <quic_amitajit@quicinc.com>